apps/api/pyproject.toml
[project]
name = "workpad-ai-api"
version = "0.1.0"
description = "FastAPI backend for the Workpad AI canvas experience."
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.40.0",
  "fastapi>=0.115.0",
  "markdown>=3.7",
  "openai>=2.0.0",
  "pydantic-settings>=2.6.1",
  "python-docx>=1.1.2",
  "reportlab>=4.2.5",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
